Refactor on_off.py
